+++ pkg/analyzer/lib/src/dart/element/type_provider.dart
@@ -11,11 +11,13 @@ import 'package:analyzer/src/dart/element/type.dart';
 const Map<String, Set<String>> _nonSubtypableClassMap = {
   'dart:async': _nonSubtypableDartAsyncClassNames,
   'dart:core': _nonSubtypableDartCoreClassNames,
+  'dart:typed_data': _nonSubtypableDartTypedDataClassNames,
 };
 
 const Set<String> _nonSubtypableClassNames = {
   ..._nonSubtypableDartCoreClassNames,
   ..._nonSubtypableDartAsyncClassNames,
+  ..._nonSubtypableDartTypedDataClassNames,
 };
 
 const Set<String> _nonSubtypableDartAsyncClassNames = {
@@ -29,6 +31,46 @@ const Set<String> _nonSubtypableDartCoreClassNames = {
   'Null',
   'num',
   'String',
+};
+
+const Set<String> _nonSubtypableDartTypedDataClassNames = {
+  'ByteBuffer',
+  'ByteData',
+  'Endian',
+  'Float32List',
+  'Float32x4',
+  'Float32x4List',
+  'Float64List',
+  'Float64x2',
+  'Float64x2List',
+  'Int16List',
+  'Int32List',
+  'Int32x4',
+  'Int32x4List',
+  'Int64List',
+  'Int8List',
+  'TypedData',
+  'Uint16List',
+  'Uint32List',
+  'Uint64List',
+  'Uint8ClampedList',
+  'Uint8List',
+  'UnmodifiableByteBufferView',
+  'UnmodifiableByteDataView',
+  'UnmodifiableFloat32ListView',
+  'UnmodifiableFloat32x4ListView',
+  'UnmodifiableFloat64ListView',
+  'UnmodifiableFloat64x2ListView',
+  'UnmodifiableInt16ListView',
+  'UnmodifiableInt32ListView',
+  'UnmodifiableInt32x4ListView',
+  'UnmodifiableInt64ListView',
+  'UnmodifiableInt8ListView',
+  'UnmodifiableUint16ListView',
+  'UnmodifiableUint32ListView',
+  'UnmodifiableUint64ListView',
+  'UnmodifiableUint8ClampedListView',
+  'UnmodifiableUint8ListView',
 };
 
 /// Provide common functionality shared by the various TypeProvider